(pc-3682): rename discovery v3 to discovery
