+++ src/components/pages/discovery/Deck/Card/__specs__/CardContainer.spec.js
