+++ src/components/pages/discovery/Deck/Navigation/__specs__/Navigation.spec.jsx
