+++ src/components/pages/discovery/Deck/Navigation/__specs__/NavigationContainer.spec.js
