+++ src/components/pages/discovery/Deck/__specs__/DeckContainer.spec.js
