+++ src/components/pages/discovery/__specs__/DiscoveryContainer.spec.js
