+++ src/components/pages/discovery/selectors/__specs__/selectCurrentRecommendation.spec.js
