+++ src/components/pages/discovery/selectors/selectCurrentRecommendation.js
