+++ src/components/router/routes.jsx
@@ -8,12 +8,13 @@ import { IcoNavProfile } from '../layout/NavBar/Icons/IcoNavProfile'
 import { IcoNavSearch } from '../layout/NavBar/Icons/IcoNavSearch'
 import ActivationContainer from '../pages/activation/ActivationContainer'
 import BetaPageContainer from '../pages/beta-page/BetaPageContainer'
-import DiscoveryContainer from '../pages/discovery/DiscoveryContainer'
 import MyBookingsContainer from '../pages/my-bookings/MyBookingsContainer'
+import DiscoveryContainer from '../pages/discovery/DiscoveryContainer'
 import MyFavoritesContainer from '../pages/my-favorites/MyFavoritesContainer'
 import ForgotPassword from '../pages/forgot-password/ForgotPassword'
 import OfferContainer from '../pages/offer/OfferContainer'
 import ProfileContainer from '../pages/profile/ProfileContainer'
+import SignupContainer from '../pages/signup/SignupContainer'
 import SearchContainer from '../pages/search/SearchContainer'
 import SignInContainer from '../pages/signin/SignInContainer'
 import TutorialsContainer from '../pages/tutorials/TutorialsContainer'
